requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
environ~=1.0
python-dotenv~=0.21.1
django-filter~=22.1